chore: move folders to track data utility folder
