+++ src/app/main/utility-components/track-data-utility/dummy-file-data.ts
@@ -1,4 +1,4 @@
-import { Files } from '../types/Files';
+import { Files } from '../../../types/Files';
 
 const randomDate = (start: Date, end: Date) => {
   return new Date(
