+++ src/app/main/utility-components/track-data-utility/track-data-utility.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AstroComponentsModule } from '@astrouxds/angular';
-import { TrackFilesComponent } from 'src/app/track-files/track-files.component';
-import { TrackDataComponent } from 'src/app/track-data/track-data.component';
+import { TrackFilesComponent } from 'src/app/main/utility-components/track-data-utility/track-files/track-files.component';
+import { TrackDataComponent } from 'src/app/main/utility-components/track-data-utility/track-data/track-data.component';
 @Component({
   selector: 'fds-track-data-util',
   standalone: true,
